backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
boto3==1.36.0
gunicorn==23.0.0
jsonschema==4.23.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-flask==1.3.0
